skill_updater/requirements.txt
streamlit>=1.40.0
asyncpg>=0.29.0
google-genai==1.68.0
openai>=1.50.0
python-dotenv>=1.0.0
pydantic>=2.6.0
